# api/openapi.yaml
openapi: 3.0.3
info:
  title: Basketball Stats Service API
  version: 1.0.0
  description: HTTP API for teams, players, games and stats, including aggregated endpoints.
servers:
  - url: /api/v1
security: []
paths:
...
  /stats:
    post:
      summary: Upsert a player's stat line for a game
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PlayerStatLineInput' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/PlayerStatLine' } } } }
        '400': { description: Invalid input, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '409': { description: Conflict (FK), content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  /games/{id}/stats:
    get:
      summary: List stat lines for a game
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/PlayerStatLine' } } } } }
        '400': { description: Invalid input, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error: { type: string, example: invalid_input }
        field_errors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }
...
    PlayerStatLineInput:
      type: object
      properties:
        player_id: { type: integer }
        game_id: { type: integer }
        points: { type: integer, minimum: 0 }
        rebounds: { type: integer, minimum: 0 }
        assists: { type: integer, minimum: 0 }
        steals: { type: integer, minimum: 0 }
        blocks: { type: integer, minimum: 0 }
        fouls: { type: integer, minimum: 0 }
        turnovers: { type: integer, minimum: 0 }
        minutes_played: { type: number, minimum: 0, maximum: 60 }
      required: [player_id, game_id]
    PlayerStatLine:
      allOf:
        - $ref: '#/components/schemas/PlayerStatLineInput'
        - type: object
          properties:
            id: { type: integer }
            created_at: { type: string, format: date-time }
            updated_at: { type: string, format: date-time }
...